requirements.txt
flask
python-dotenv
PyJWT[crypto]
requests
